+++ packages/flexi_cart/lib/src/cart_cache/cart_cache.dart
@@ -1,0 +1,2 @@
+export 'cart_cache_provider.dart';
+export 'flexi_cart_persistence.dart';